Split engine into modules
